Laba 6
Two mixers were added: one for the music, a long background composition, and the second for short sounds

The sounds for the character's jump and movement were added, with the specific sounds chosen as desired. The sounds were routed through the sound mixer.

A menu was added, which could be opened by pressing the Esc key, pausing the game, and featuring buttons to return to the game or exit the game. Additionally, sliders were included to adjust the volume of the sounds in the two mixers: one slider for the music mixer volume and the second for the sound mixer volume.