Laba 8,9
Lab 8.
Added HP, which decreased when colliding with traps and enemies. HP was implemented either as a progress bar that shrank or as game objects, the number of which was equivalent to the HP value.
When HP reached 0, the character was teleported back to the last checkpoint.
When falling into a pit, HP was reset and recreated at full value, and the character was teleported to the checkpoint.

Lab 9.
Added an enemy with running animation and logic. The enemy could move either within a certain distance or using colliders.